+++ src/components/Popup.js
@@ -33,52 +33,4 @@ class Popup {
   }
 }
 
-class PopupWithImage extends Popup {
-    constructor(popupSelector) {
-      super(popupSelector);
-      this._bigImage = this._popupSelector.querySelector('.popup__big-img');
-      this._popupImgDesc = this._popupSelector.querySelector('.popup__img-description');
-    }
-
-    open(item) {
-      super.open();
-      this._bigImage.src = item.link;
-      this._bigImage.alt = item.name;
-      this._popupImgDesc.textContent = item.name;
-    }
-}
-
-class PopupWithForm extends Popup {
-  constructor(popupSelector, handler) {
-    super(popupSelector);
-    this._handler = handler;
-    this._formElement = this._popupSelector.querySelector('.popup__form');
-    this._inputList = Array.from(this._formElement.querySelectorAll('.popup__input-field'));
-    this._btnSubmit = this._popupSelector.querySelector('.popup__submit-button');
-  }
-
-  _getInputValues() {
-    this._formValues = {};
-    this._inputList.forEach(input => {
-      this._formValues[input.name] = input.value;
-    })
-    return this._formValues
-  }
-
-  close() {
-    super.close();
-    this._formElement.reset();
-  }
-
-  setEventListeners() {
-    super.setEventListeners();
-    console.log(this);
-    this._formElement.addEventListener('submit', (e) => {
-      e.preventDefault();
-      const formData = this._getInputValues();
-      this._handler.formSubmitHandler(formData,btnSubmit);
-    })
-  }
-}
-
-export { Popup, PopupWithImage, PopupWithForm }
+export { Popup }
