+++ src/components/card.js
@@ -13,8 +13,8 @@ class Card {
   }
 
   _getTemplate() {
-    const cardElement = document.querySelector(this._templateSelector).content.querySelector('.card').cloneNode(true);
-    return cardElement;
+    this._cardElement = document.querySelector(this._templateSelector).content.querySelector('.card').cloneNode(true);
+    return this._cardElement;
   }
 
   generate() {
